move the waveform component to utils for better separation of concerns
